chore(tauri/ui): move components into seperate dir
