+++ tauri-app/src/lib/components/BadgeActive.svelte
