+++ tauri-app/src/lib/components/BadgeExternalLink.svelte
