+++ tauri-app/src/lib/components/ButtonLoading.svelte
