+++ tauri-app/src/lib/components/IconExternalLink.svelte
