+++ tauri-app/src/lib/components/InputHex.svelte
