+++ tauri-app/src/lib/components/InputToken.svelte
