+++ tauri-app/src/lib/components/InputTokenAmount.svelte
