+++ tauri-app/src/lib/components/ModalVaultDepositGeneric.svelte
@@ -1,15 +1,14 @@
 <script lang="ts">
   import { Button, Modal, Label } from 'flowbite-svelte';
-  import InputTokenAmount from '$lib/InputTokenAmount.svelte';
-  import { vaultWithdraw } from '$lib/utils/vaultWithdraw';
-  import InputToken from '$lib/InputToken.svelte';
-  import InputVaultId from './InputVaultId.svelte';
-  import ButtonLoading from '$lib/ButtonLoading.svelte';
+  import InputTokenAmount from '$lib/components/InputTokenAmount.svelte';
+  import { vaultDeposit } from '$lib/utils/vaultDeposit';
+  import InputToken from '$lib/components/InputToken.svelte';
+  import InputVaultId from '$lib/components/InputVaultId.svelte';
+  import ButtonLoading from '$lib/components/ButtonLoading.svelte';
 
   export let open = false;
-
-  let vaultId: bigint;
-  let tokenAddress: string;
+  let vaultId: bigint = 0n;
+  let tokenAddress: string = '';
   let tokenDecimals: number = 0;
   let amount: bigint;
   let isSubmitting = false;
@@ -26,7 +25,7 @@
   async function execute() {
     isSubmitting = true;
     try {
-      await vaultWithdraw(vaultId, tokenAddress, amount);
+      await vaultDeposit(vaultId, tokenAddress, amount);
       reset();
       // eslint-disable-next-line no-empty
     } catch (e) {}
@@ -34,7 +33,7 @@
   }
 </script>
 
-<Modal title="Withdraw from Vault" bind:open outsideclose size="sm" on:close={reset}>
+<Modal title="Deposit to Vault" bind:open outsideclose size="sm" on:close={reset}>
   <div>
     <h5 class="mb-2 w-full text-xl font-bold tracking-tight text-gray-900 dark:text-white">
       Vault ID
@@ -54,7 +53,7 @@
       for="amount"
       class="mb-2 w-full text-xl font-bold tracking-tight text-gray-900 dark:text-white"
     >
-      Target Amount
+      Amount
     </Label>
     <InputTokenAmount bind:value={amount} decimals={tokenDecimals} />
   </div>
@@ -63,7 +62,7 @@
     <div class="flex w-full justify-end space-x-4">
       <Button color="alternative" on:click={reset} disabled={isSubmitting}>Cancel</Button>
       <ButtonLoading on:click={execute} disabled={!amount || amount === 0n || isSubmitting} loading={isSubmitting}>
-        Make Withdrawal
+        Submit Deposit
       </ButtonLoading>
     </div>
   </svelte:fragment>
