+++ tauri-app/src/lib/components/ModalVaultWithdraw.svelte
