+++ tauri-app/src/lib/components/ModalVaultWithdrawGeneric.svelte
@@ -1,14 +1,15 @@
 <script lang="ts">
   import { Button, Modal, Label } from 'flowbite-svelte';
   import InputTokenAmount from '$lib/InputTokenAmount.svelte';
-  import { vaultDeposit } from '$lib/utils/vaultDeposit';
-  import InputToken from '$lib/InputToken.svelte';
-  import InputVaultId from '$lib/InputVaultId.svelte';
-  import ButtonLoading from '$lib/ButtonLoading.svelte';
+  import { vaultWithdraw } from '$lib/utils/vaultWithdraw';
+  import InputToken from '$lib/components/InputToken.svelte';
+  import InputVaultId from '$lib/components/InputVaultId.svelte';
+  import ButtonLoading from '$lib/components/ButtonLoading.svelte';
 
   export let open = false;
-  let vaultId: bigint = 0n;
-  let tokenAddress: string = '';
+
+  let vaultId: bigint;
+  let tokenAddress: string;
   let tokenDecimals: number = 0;
   let amount: bigint;
   let isSubmitting = false;
@@ -25,7 +26,7 @@
   async function execute() {
     isSubmitting = true;
     try {
-      await vaultDeposit(vaultId, tokenAddress, amount);
+      await vaultWithdraw(vaultId, tokenAddress, amount);
       reset();
       // eslint-disable-next-line no-empty
     } catch (e) {}
@@ -33,7 +34,7 @@
   }
 </script>
 
-<Modal title="Deposit to Vault" bind:open outsideclose size="sm" on:close={reset}>
+<Modal title="Withdraw from Vault" bind:open outsideclose size="sm" on:close={reset}>
   <div>
     <h5 class="mb-2 w-full text-xl font-bold tracking-tight text-gray-900 dark:text-white">
       Vault ID
@@ -53,7 +54,7 @@
       for="amount"
       class="mb-2 w-full text-xl font-bold tracking-tight text-gray-900 dark:text-white"
     >
-      Amount
+      Target Amount
     </Label>
     <InputTokenAmount bind:value={amount} decimals={tokenDecimals} />
   </div>
@@ -62,7 +63,7 @@
     <div class="flex w-full justify-end space-x-4">
       <Button color="alternative" on:click={reset} disabled={isSubmitting}>Cancel</Button>
       <ButtonLoading on:click={execute} disabled={!amount || amount === 0n || isSubmitting} loading={isSubmitting}>
-        Submit Deposit
+        Make Withdrawal
       </ButtonLoading>
     </div>
   </svelte:fragment>
